# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.ChangePasswordRequest:
    properties:
      confirm_new_password:
        type: string
      current_password:
        type: string
      new_password:
        minLength: 6
        type: string
    required:
    - confirm_new_password
    - current_password
    - new_password
    type: object
...
  models.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  models.RegisterRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        minLength: 6
        type: string
      username:
        type: string
    required:
    - email
    - full_name
    - password
    - username
    type: object
...
  models.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      role:
        type: string
      username:
        type: string
    type: object
  utils.Meta:
    properties:
      filters:
        additionalProperties: true
        type: object
      pagination:
        $ref: '#/definitions/utils.Pagination'
    type: object
  utils.PaginatedResponse:
    properties:
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/utils.Meta'
      status:
        type: integer
    type: object
  utils.Pagination:
    properties:
      current_page:
        type: integer
      has_next:
        type: boolean
      has_prev:
        type: boolean
      items_per_page:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  utils.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API backend cho ứng dụng shop bán hàng. Hỗ trợ quản lý sản phẩm,
    xác thực người dùng và quản trị hệ thống.
  title: Backend Shop API
  version: "1.0"
paths:
  /admin/users:
    get:
      description: Lấy danh sách tất cả người dùng (chỉ admin)
      parameters:
      - default: 1
        description: Trang hiện tại
        in: query
        name: page
        type: integer
      - default: 10
        description: Số user mỗi trang
        in: query
        maximum: 100
        name: limit
        type: integer
      - description: Tìm kiếm theo username/email
        in: query
        name: search
        type: string
      - description: Lọc theo role
        enum:
        - admin
        - user
        in: query
        name: role
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.UserResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Lấy danh sách người dùng
      tags:
      - Admin
  /auth/login:
    post:
      consumes:
      - application/json
      description: Đăng nhập bằng username và password, trả về JWT token
      parameters:
      - description: Thông tin đăng nhập
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Đăng nhập
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Tạo tài khoản mới với username, email và password
      parameters:
      - description: Thông tin đăng ký
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Đăng ký tài khoản
      tags:
      - Auth
...
  /users/change-password:
    put:
      consumes:
      - application/json
      description: Đổi mật khẩu người dùng hiện tại
      parameters:
      - description: Thông tin đổi mật khẩu
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Đổi mật khẩu
      tags:
      - User
securityDefinitions:
  BearerAuth:
    description: 'Nhập token theo format: Bearer {access_token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"