# internal/handlers/admin_handler.go
package handlers

import (
	"net/http"

	"github.com/NgTruong624/project_backend/internal/models"
	"github.com/NgTruong624/project_backend/internal/repository"
	"github.com/NgTruong624/project_backend/internal/utils"
	"github.com/gin-gonic/gin"
	"gorm.io/gorm"
)

type AdminHandler struct {
	userRepo *repository.UserRepository
}

func NewAdminHandler(db *gorm.DB) *AdminHandler {
	return &AdminHandler{
		userRepo: repository.NewUserRepository(db),
	}
}

// GetUsersList lấy danh sách tất cả người dùng (Admin only)
// GetUsersList godoc
// @Summary      Lấy danh sách người dùng
// @Description  Lấy danh sách tất cả người dùng (chỉ admin)
// @Tags         Admin
// @Produce      json
// @Param        page   query int    false "Trang hiện tại" default(1)
// @Param        limit  query int    false "Số user mỗi trang" default(10) maximum(100)
// @Param        search query string false "Tìm kiếm theo username/email"
// @Param        role   query string false "Lọc theo role" Enums(admin, user)
// @Success      200 {object} utils.PaginatedResponse{data=[]models.UserResponse}
// @Failure      400 {object} utils.Response
// @Failure      403 {object} utils.Response
// @Security     BearerAuth
// @Router       /admin/users [get]
func (h *AdminHandler) GetUsersList(c *gin.Context) {
	// Kiểm tra quyền admin
	role := c.GetString("role")
	if role != "admin" {
		c.JSON(http.StatusForbidden, utils.NewErrorResponse(http.StatusForbidden, "Permission denied", "Only admin can access user list"))
		return
	}

	var query models.UserQueryParams
	if err := c.ShouldBindQuery(&query); err != nil {
		c.JSON(http.StatusBadRequest, utils.NewErrorResponse(http.StatusBadRequest, "Invalid query parameters", err.Error()))
		return
	}

	// Set default values
	if query.Page <= 0 {
		query.Page = 1
	}
	if query.Limit <= 0 {
		query.Limit = 10
	}
	if query.Limit > 100 {
		query.Limit = 100
	}

	// Get users from repository
	users, total, err := h.userRepo.GetAllUsers(&query)
	if err != nil {
		c.JSON(http.StatusInternalServerError, utils.NewErrorResponse(http.StatusInternalServerError, "Error fetching users", err.Error()))
		return
	}

	// Convert to response (remove password field)
	var userResponses []models.UserResponse
	for _, u := range users {
		userResponses = append(userResponses, models.UserResponse{
			ID:        u.ID,
			Username:  u.Username,
			Email:     u.Email,
			FullName:  u.FullName,
			Role:      u.Role,
			CreatedAt: u.CreatedAt,
		})
	}

	// Calculate pagination info
	totalPages := (int(total) + query.Limit - 1) / query.Limit

	// Prepare metadata
	meta := map[string]interface{}{
		"total":        total,
		"total_pages":  totalPages,
		"current_page": query.Page,
		"per_page":     query.Limit,
		"has_next":     query.Page < totalPages,
		"has_prev":     query.Page > 1,
	}

	// Add filter info to metadata
	if query.Search != "" {
		meta["search"] = query.Search
	}
	if query.Role != "" {
		meta["role"] = query.Role
	}

	c.JSON(http.StatusOK, utils.NewPaginatedResponse(
		http.StatusOK,
		"Users retrieved successfully",
		userResponses,
		query.Page,
		totalPages,
		total,
		query.Limit,
		meta,
	))
}